requirements.txt
dashscope
numpy
pypinyin
python-dotenv
requests
sounddevice
